LayoutTests/media/modern-media-controls/button/button-focus-state-expected.txt

Testing an IconButton has a blue background-color when focused.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Confirm inital icon button color.
PASS window.getComputedStyle(button.image.element).backgroundColor is "rgba(255, 255, 255, 0.55)"
Confirm icon button color changed after focus.
PASS window.getComputedStyle(button.image.element).backgroundColor is not "rgba(255, 255, 255, 0.55)"
PASS successfullyParsed is true

TEST COMPLETE
